增加热度统计

+++ index.js
@@ -24,28 +24,49 @@ p2p.ignore(function (infohash, rinfo, callback) {
 });
 
 p2p.on('metadata', function (metadata) {
-    var torrentFilePathSaveTo = path.join(__dirname, "torrents", metadata.infohash + ".torrent");
+    var time = moment().format('YYYY-MM-DD');
+    var torrentDir = path.join(__dirname, "torrents/" + time);
+    var torrentFilePathSaveTo = path.join(torrentDir, metadata.infohash + ".torrent");
     var sql = "insert into t_torrent(id,name,type,size,files,create_date,enable,magnet,torrent_path) values (?,?,?,?,?,?,?,?,?)";
     var data = torrent(metadata.info);
     var now = moment().format('YYYY-MM-DD HH:mm:ss');
     var magnet = 'magnet:?xt=urn:btih:' + metadata.infohash;
     var files = JSON.stringify(data.files);
     files = files.replace(/,/g,"\,");
-    var time = moment().format('YYYY-MM-DD');
     var torrentPath = "/torrents/" + time + "/" + metadata.infohash + ".torrent";
     var sqlParam = [metadata.infohash,data.name,data.type,data.size,files,now,1,magnet,torrentPath];
     db.insert(sql,sqlParam,function(err,result) {
         if(err) {
             console.error(err);
+        } else {
+            console.log("insert torrent result:" + result);
+            fs.exists(torrentDir,function(exist) {
+                if(!exist) {
+                    fs.mkdir(torrentDir,function(err) {
+                        if(err) {
+                            console.error(err);
+                        } else {
+                            fs.writeFile(torrentFilePathSaveTo, bencode.encode({'info': metadata.info}), function(err) {
+                                if (err) {
+                                    return console.error(err);
+                                }
+                                console.log(metadata.infohash + ".torrent has saved.");
+                            });
+                        }
+                    });
+                } else {
+                    fs.writeFile(torrentFilePathSaveTo, bencode.encode({'info': metadata.info}), function(err) {
+                        if (err) {
+                            return console.error(err);
+                        }
+                        console.log(metadata.infohash + ".torrent has saved.");
+                    });
+                }
+            });
         }
-        console.log("insert torrent result:" + result);
+        
     });
-    fs.writeFile(torrentFilePathSaveTo, bencode.encode({'info': metadata.info}), function(err) {
-        if (err) {
-            return console.error(err);
-        }
-        console.log(metadata.infohash + ".torrent has saved.");
-    });
+    
 });
 
 p2p.listen(6881, '0.0.0.0');
